Msg allocator now sets msg->pl_size even when message comes from free_list
Now msg_allocator_alloc correctly sets msg->pl_size to payload_size even
when the message is popped from the free_list. Without this, it would
cause a bug in handling of pubsub messages because of a mismatch between
the msg->pl_size and the actual size of the payload.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>