Experimental branch for debugging
Running with MPI using pubsub module yields different results than running without.
This branch is used to better explore this issue and will likely be destroyed after.

To more consistently halt simulation, the processing loop is modified to simply
ignore messages scheduled for times coming after the desired end time.
This modification is only for the sake of consistency in results and by no means was
demonstrated to be safe to implement otherwise, even if one was to ignore the
memory leaks the current implementation of this approach obviously presents.

All other modifications have been made for the same reasons.

The only actual fix is skipping the reading of the neural probe when
silent processing after a rollback.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>