PubSub delivered messages are processed rather than being inserted in the queue
Slightly changed PubSub paradigm: delivered messages (at LP level) are no longer
put in queue, but rather processed directly.

This appears to bring about better performance, due to the decrease of both
queue size and number of insertions/extractions from the queue, even in
spite of the increased rollback likelihood that processing directly might cause.

The new function PubsubDeliver is added, that the user needs to call from
within ProcessPublishedEvent to provide the specifics of the delivered message,
in place of calling ScheduleNewEvent.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>