Added new initialization membrane potential to cortical microcircuit model
In cortical microcircuit model (currently Shimoura18 - to be changed
in Potjans14), a new way to initialize the membrane potential was added:
the user can decide to use either the original values, with one mean and
standard deviation for all populations, or the optimized values, where
each population has a specific mean and standard deviation. The optimized
version was taken from NEST simulator's pyNest implementation of
Potjans_2014 - where it is used by default - and allows a reduction of
the initial activity burst in the network.

Also only one neuron gets probed now, instead of all.

Reduced Buddy allocator block size to make allocations finer, as well
as the amount of memory dedicated to a single LP, as neurons require very
small amounts of memory each.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>