Add subbed node is not called when subscriber is local
Since we rely on having the bit regarding local node not set in the
subscriber nodes bitmap, we now do not try to set it if subscriber is local.

When receiving a pubsub antimessage via MPI, the PUBSUB flag is only
set to propagate it into the children messages. It is now reset before
freeing, to free the node-level antimessage as a normal message.

Simulation now gets to start but we now have a heap buffer overflow when
populating the published node-level message.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>