Fixed heap buffer overflow
Heap buffer overflow was caused by bad calculation of pub_info_ptr-s.
They are now computed correctly.

Added a check in pubsub_insert_in_past for the edge case in which the
msg has to be inserted at the start of the past_pubsubs array.

Canged sub_node_handle_published_message to have an assert on children
count, rather than a normal check. Since LPs should only send messages
to nodes that manage at least one LP subscribed to them, if a node
receives a pubsub message from an LP to which no local LP is subbed
it is a symptom of a malformed topology, or an internal bug.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>