Added facility to log committed published messages to file
To better study the execution, it is heplful to have a trace of
the messages each LP has published. For this purpose a function was
created to log them. The logging is called from the fossil collection
and the cleanup functions. This way only committed messages are logged.

Bug: It seems that some messages are duplicated in the logs, probably
caused by a small oversight.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>