Changed pubsub internals on delivering
Added new function PubsubDeliver that the user has to call in response
to HandlePublishedMessage, to provide the result of the unpacking.
This allowed us to remove the dirty interaction with ScheduleNewEvent
that had to take place before.

process_msg now takes a msg pointer as argument, rather than
extracting it by itself.
New function process_next_msg is introduced, that extracts the
next message and calls process_msg with it.

Fixed the sanity check in ScheduleRetractableEvent that would crash
in silent execution.

Signed-off-by: AdrianoPi <24545691+AdrianoPi@users.noreply.github.com>